refactor: implement Phase 1 of ishortcut_action.py modularization
- Create specialized modules for UI event handlers per issue #39
- Migrate key functions to model_info_actions, download_actions, etc.
- Maintain backward compatibility through main ishortcut_action.py
- All tests passing with refactored structure
- Functions like on_update_information_btn_click properly migrated

Resolves #39 (Phase 1)
